Log: Implemented project compression + logging
- Removed CodeStateID as a mandatory field
- Moved compression + logging to service
- Removed CodeStateID as a mandatory field (from the service only)
- Began testing
