+++ PatternPal/PatternPal.LoggingServer/Protos/logcollector.proto
@@ -19,7 +19,6 @@ EventType event_type = 1;
 string event_id = 2;
 string subject_id = 3;
 string tool_instances = 4;
-string code_state_id = 5;
 
 // Optional columns
 string client_timestamp = 6;
@@ -42,7 +41,6 @@ string intervention_category = 22;
 string intervention_type = 23;
 string intervention_message = 24;
 optional bytes data = 25;
-
 }
 
 // Input arguments for recognizer response
