+++ PatternPal/PatternPal.Tests/LoggingService/Resources/OnlyCsFiles/Hammer.cs
