+++ PatternPal/PatternPal.Tests/LoggingService/Resources/OnlyCsFiles/IEnchantment.cs
