+++ PatternPal/PatternPal.Tests/LoggingService/Resources/OnlyCsFiles/IWeapon.cs
