+++ PatternPal/PatternPal.Tests/LoggingService/Resources/OnlyCsFiles/SoulEatingEnchantment.cs
