+++ PatternPal/PatternPal.Tests/LoggingService/Resources/OnlyCsFiles/Sword.cs
