+++ PatternPal/PatternPal/Services/LoggingService.cs
@@ -1,5 +1,7 @@
-#region
+﻿#region
 
+using Google.Protobuf;
+using System.IO.Compression;
 using Grpc.Net.Client;
 using PatternPal.LoggingServer;
 using ExecutionResult = PatternPal.LoggingServer.ExecutionResult;
@@ -14,6 +16,17 @@ namespace PatternPal.Services;
  */
 public class LoggingService : LogProviderService.LogProviderServiceBase
 {
+    /// <summary>
+    /// Stores the last codeState that has been successfully logged to the server, per projectId.
+    /// </summary>
+    private static Dictionary<String, ByteString> _lastCodeState = new Dictionary<string, ByteString>();
+
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest">TODO</param>
+    /// <param name="context">TODO</param>
+    /// <returns></returns>
     public override Task<LogEventResponse> LogEvent(LogEventRequest receivedRequest, ServerCallContext context)
     {
         GrpcChannel grpcChannel = GrpcChannel.ForAddress(
@@ -21,7 +34,15 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
 
         LogRequest sendRequest = DetermineSpecificLog(receivedRequest);
         LogCollectorService.LogCollectorServiceClient client = new(grpcChannel);
-        client.Log(sendRequest);
+
+        // TODO What should be done with the actual response of the logging server?
+        LogResponse res = client.Log(sendRequest); 
+
+        // TODO First check if logging was successful
+        if (sendRequest.HasData)
+        {
+            _lastCodeState.Add(sendRequest.ProjectId, sendRequest.Data);
+        }
 
         // Send response back to frond-end to verify something has been received here
         LogEventResponse response = new() { ResponseMessage = "a response message." };
@@ -46,7 +67,6 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
         };
     }
 
-
     /// <summary>
     /// Each log has certain required fields that are independent of the EventType. These fields are set here.
     /// </summary>
@@ -59,7 +79,6 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
             EventId = receivedRequest.EventId, //TODO: Generate ID in Logging Server self
             SubjectId = receivedRequest.SubjectId,
             ToolInstances = Environment.Version.ToString(),
-            CodeStateId = Guid.NewGuid().ToString(), //TODO: Generate ID in Logging Server
             ClientTimestamp =
                 DateTime.UtcNow.ToString(
                     "yyyy-MM-dd HH:mm:ss.fff zzz"), //TODO: DateTimeOffset.Now.ToString("yyyy-MM-dd HH:mm:ss.fff  zzz"), : Logging server cannot work with offsets yet
@@ -69,6 +88,11 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
     }
 
     #region Log types
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest CompileLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
@@ -77,6 +101,11 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
         return sendLog;
     }
 
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest CompileErrorLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
@@ -88,24 +117,50 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
         sendLog.SourceLocation = receivedRequest.SourceLocation;
         return sendLog;
     }
+
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest ProjectCloseLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
         sendLog.EventType = LoggingServer.EventType.EvtProjectClose;
         sendLog.ProjectId = receivedRequest.ProjectId;
+
+        if (receivedRequest.HasFilePath)
+        {
+            sendLog.Data = ZipDirectory(receivedRequest.FilePath);
+        }
+
         return sendLog;
     }
 
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest ProjectOpenLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
         sendLog.EventType = LoggingServer.EventType.EvtProjectOpen;
         sendLog.ProjectId = receivedRequest.ProjectId;
-        sendLog.Data = receivedRequest.Data;
+
+        if (receivedRequest.HasFilePath)
+        {
+            sendLog.Data = ZipDirectory(receivedRequest.FilePath);
+        }
 
         return sendLog;
     }
 
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest DebugProgramLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
@@ -115,6 +170,12 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
 
         return sendLog;
     }
+
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest SessionStartLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
@@ -122,12 +183,76 @@ public class LoggingService : LogProviderService.LogProviderServiceBase
 
         return sendLog;
     }
+
+    /// <summary>
+    /// TODO
+    /// </summary>
+    /// <param name="receivedRequest"></param>
+    /// <returns></returns>
     private static LogRequest SessionEndLog(LogEventRequest receivedRequest)
     {
         LogRequest sendLog = StandardLog(receivedRequest);
         sendLog.EventType = LoggingServer.EventType.EvtSessionEnd;
       
         return sendLog;
+    }
+    #endregion
+
+    #region Utils
+    /// <summary>
+    ///  TODO
+    /// </summary>
+    /// <param name="path"></param>
+    /// <returns></returns>
+    public static ByteString ZipDirectory(string path)
+    {
+        // TODO: Protect against too large codebases.
+        // TODO: Now also includes *.cs-build artifacts.
+        Byte[] bytes;
+
+        using (MemoryStream ms = new MemoryStream())
+        {
+            using (ZipArchive archive = new ZipArchive(ms, ZipArchiveMode.Create))
+            {
+
+                string[] files = Directory.GetFiles(path, "*.cs", SearchOption.AllDirectories);
+
+                foreach (string file in files)
+                {
+                    ZipArchiveEntry entry = archive.CreateEntry(GetRelativePath(path, file), CompressionLevel.Optimal);
+
+                    using (Stream entryStream = entry.Open())
+                    using (FileStream contents = File.OpenRead(file))
+                    {
+                        contents.CopyTo(entryStream);
+                    }
+                }
+            }
+
+            bytes = ms.ToArray();
+        }
+
+        return ByteString.CopyFrom(bytes);
+    }
+
+    // TODO Separate utility because of duplication with extension
+    /// <summary>
+    /// Gets the relative path when given an absolute directory path and a filename.
+    /// </summary>
+    /// <param name="relativeTo"> The absolute path to the root folder</param>
+    /// <param name="path">The absolute path to the specific file</param>
+    /// <returns></returns>
+    public static string GetRelativePath(string relativeTo, string path)
+    {
+        Uri uri = new Uri(relativeTo);
+        string rel = Uri.UnescapeDataString(uri.MakeRelativeUri(new Uri(path)).ToString())
+            .Replace(Path.AltDirectorySeparatorChar, Path.DirectorySeparatorChar);
+        if (rel.Contains(Path.DirectorySeparatorChar.ToString()) == false)
+        {
+            rel = $".{Path.DirectorySeparatorChar}{rel}";
+        }
+
+        return rel;
     }
     #endregion
 }
